signal-bot-docker/signal-bot/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pandas==2.2.3
numpy==2.1.3
python-dotenv==1.0.1
structlog==24.4.0
